+++ NEWS.md
@@ -43,6 +43,7 @@
   * Xbox360_XEX: Xbox 360 executables.
   * BRSTM: Nintendo Wii BRSTM audio files.
   * BCSTM: Nintendo 3DS BCSTM audio files.
+    * Also supports Nintendo Wii U BFSTM audio files.
 
 * New audio parsers:
   * ADX: CRI ADX audio format. Used by many Sega games.